Optimize ConformerEncoder performance and fix gradient computation
- Optimize apply_channel_mask: use broadcasting instead of explicit expand
- Optimize _create_masks: use expand instead of repeat to avoid memory copy
- Fix inplace operation issue in apply_channel_mask that caused gradient computation error
- Disable dither in preprocessor for deterministic alignment
- Add conv_split_by_channel method for better performance on large inputs
- Update config to disable dither for alignment testing

+++ nest_ssl_project/modules/ssl_modules/quantizers.py
@@ -69,7 +69,8 @@ class RandomProjectionVectorQuantizer(NeuralModule, Exportable):
         torch.nn.init.xavier_normal_(self.proj.weight)
 
         # (num_books, num_classes, hid_dim)
-        codebooks = torch.randn(self.num_books, self.num_classes, self.code_dim, dtype=torch.float32)
+        # Use double() to match NeMo's implementation (float64)
+        codebooks = torch.randn(self.num_books, self.num_classes, self.code_dim).double()
         torch.nn.init.normal_(codebooks, mean=0, std=1)
         codebooks = F.normalize(codebooks, dim=-1)
         self.codebooks = nn.Parameter(codebooks)
